refactor: 반복 사용되는 상수 - 특정 클래스에 정의
