+++ src/main/java/lotto/domain/LottoTicket.java
@@ -2,10 +2,9 @@ package lotto.domain;
 
 import java.util.ArrayList;
 
+import static lotto.utils.Constants.*;
+
 public class LottoTicket {
-    public static final int LOTTO_TICKET_SIZE = 6;
-    public static final int LOTTO_LOWER_BOUND = 1;
-    public static final int LOTTO_UPPER_BOUND = 45;
     private final ArrayList<Integer> lottoNumbers;
 
     public LottoTicket(ArrayList<Integer> lottoNumbers) {
