+++ src/main/java/lotto/utils/Constants.java
@@ -1,0 +1,9 @@
+package lotto.utils;
+
+public final class Constants {
+    public static final int LOTTO_TICKET_SIZE = 6;
+    public static final int LOTTO_LOWER_BOUND = 1;
+    public static final int LOTTO_UPPER_BOUND = 45;
+    public static final int MIN_PURCHASE_PRICE = 1000;
+
+}
